pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pom"
version = "0.0.1"
description = "Pomegranate omni speech model experiments (Qwen3 0.6B backbone)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.8.*",
    "torchaudio==2.8.*",
    "transformers>=4.41.0",
    "fastapi",
    "uvicorn",
    "websockets",
    "diffusers",
    "huggingface-hub>=0.23.0",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "einops>=0.7.0",
    "soundfile>=0.13.1",
]

[project.optional-dependencies]
train = [
    "accelerate>=0.30.0",
    "onnxruntime",
    "wandb>=0.16.0",
]
dev = ["pytest>=8.2.0"]

[tool.setuptools.packages.find]
namespaces = true
include = ["model*", "cosyvoice2*", "inference*", "train*", "app*"]
exclude = ["utils*", "reference*", "data*"]

[tool.setuptools.package-data]
app = ["*.html", "*.js", "*.css"]

[tool.pytest.ini_options]
tmp_path_retention_count = "1"
tmp_path_retention_policy = "failed"

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
